+++ apps/indexer/lib/indexer/block/catchup/fetcher.ex
@@ -185,6 +185,7 @@ defmodule Indexer.Block.Catchup.Fetcher do
       {:ok, %{inserted: inserted, errors: errors}} ->
         errors = cap_seq(sequence, errors)
         retry(sequence, errors)
+        clear_missing_ranges(range, errors)
 
         {:ok, inserted: inserted}
 
@@ -270,6 +271,14 @@ defmodule Indexer.Block.Catchup.Fetcher do
     |> Enum.map(&push_back(sequence, &1))
   end
 
+  defp clear_missing_ranges(initial_range, errors) do
+    success_numbers = Enum.to_list(initial_range) -- Enum.map(errors, &block_error_to_number/1)
+
+    success_numbers
+    |> numbers_to_ranges()
+    |> Enum.map(&MissingBlockRange.delete_range/1)
+  end
+
   defp block_errors_to_block_number_ranges(block_errors) when is_list(block_errors) do
     block_errors
     |> Enum.map(&block_error_to_number/1)
@@ -295,7 +304,7 @@ defmodule Indexer.Block.Catchup.Fetcher do
         number, range ->
           {:cont, range, number..number}
       end,
-      fn range -> {:cont, range} end
+      fn range -> {:cont, range, nil} end
     )
   end