+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [#7355](https://github.com/blockscout/blockscout/pull/7355) - Add endpoint for token info import
 - [#7393](https://github.com/blockscout/blockscout/pull/7393) - Realtime fetcher max gap
 - [#7436](https://github.com/blockscout/blockscout/pull/7436) - TokenBalanceOnDemand ERC-1155 support
+- [#7469](https://github.com/blockscout/blockscout/pull/7469) - Clear missing block ranges after every success import
 
 ### Fixes